login and register work with token as a response
