+++ src/dummycomps/Registration2.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const registration = () => {
+  return <div>registration page</div>;
+};
+
+export default registration;
